Update README.md
removed AI generated response. I went through a momentary decision as I was first pasting and left it so as not to mislead anyone that the content was AI assisted, but now, I feel that's become an universal assumption. I used AI to help organize and reword this content. It also required a great deal of effort because there has never been a Has-Needs before and blockchains aren't used the way I want to, SOoo all that is to say the AI was useless for independent content because it was trained on all my material.